requirements.txt
Flask==1.0.2
gunicorn==19.9.0
python-dotenv==0.9.1
requests==2.19.1
smart-getenv==1.1.0
Unidecode==1.0.22
